+++ ChangeLog.md
@@ -5,11 +5,12 @@
 - Distributed computation of persistent homology! (IEEE TPDS 2025)
 - Separatrix stability, a.k.a., BondMatcher (IEEE TVCG 2025)
 - Discrete vector field topology! (IEEE VIS 2024)
+- Persistence Diagram Dictionaries (IEEE TVCG 2024)
 - Fast DelaunayRips persistence computation for point cloud data
 - New backend for TrackingFromFields (critical point based)
 - Stochastic discrete gradient (IEEE VIS 2012)
 - Fast planar Rips filtration persistence computation
-- Migration to ParaView 6
+- Migration to ParaView 6.1
 - CI updates
 
 ### 1.3.0